Issue #131 chore: rename remote directory
