--- app/src/main/java/com/gdsc/sogongsogong/data/remote/Api.kt
@@ -1,8 +1,0 @@
-package com.gdsc.sogongsogong.data.remote
-
-import retrofit2.http.GET
-
-interface Api {
-    @GET()
-    suspend fun fetch(): String
-}
