+++ app/src/main/java/com/gdsc/sogongsogong/di/module/DataSourceModule.kt
@@ -1,7 +1,7 @@
 package com.gdsc.sogongsogong.di.module
 
-import com.gdsc.sogongsogong.data.remote.datasource.PostRemoteDataSource
-import com.gdsc.sogongsogong.data.remote.datasource.PostRemoteDataSourceImpl
+import com.gdsc.sogongsogong.data.remote.PostRemoteDataSource
+import com.gdsc.sogongsogong.data.remote.PostRemoteDataSourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
