+++ app/src/test/java/com/gdsc/sogongsogong/ui/board/BoardViewModelTest.kt
@@ -1,7 +1,7 @@
 package com.gdsc.sogongsogong.ui.board
 
 import com.gdsc.sogongsogong.FakeDispatcherProvider
-import com.gdsc.sogongsogong.data.remote.datasource.PostRemoteDataSource
+import com.gdsc.sogongsogong.data.remote.PostRemoteDataSource
 import io.kotest.core.spec.Spec
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
